Extract content dump loader to a separate class
